migrated Content component files to .tsx and updated snapshot
